Even more aggressive decrementing seems to work unless LNodes are forced with Hash_Mod;

+++ PerformanceTests/PerformanceTests.cpp
@@ -203,7 +203,7 @@ class Ctrie_Job : public Zeni::Concurrency::Job {
   };
 
 public:
-  Ctrie_Job(const std::shared_ptr<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint32_t>> hash_trie_, const size_t num_successors_) : hash_trie(hash_trie_), num_successors(num_successors_) {}
+  Ctrie_Job(const std::shared_ptr<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint8_t>> hash_trie_, const size_t num_successors_) : hash_trie(hash_trie_), num_successors(num_successors_) {}
 
   void execute() noexcept {
     for (int i = 0; i != 100; ++i) {
@@ -219,13 +219,13 @@ public:
       get_Job_Queue()->give_one(std::make_shared<Ctrie_Job>(hash_trie, num_successors - 1));
   }
 
-  const std::shared_ptr<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint32_t>> hash_trie;
+  const std::shared_ptr<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint8_t>> hash_trie;
   const size_t num_successors;
 };
 
 void test_Ctrie_Performance(const std::shared_ptr<Zeni::Concurrency::Worker_Threads> &worker_threads, const std::shared_ptr<Zeni::Concurrency::Job_Queue> &job_queue)
 {
-  const auto hash_trie = std::allocate_shared<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint32_t>>(Zeni::Concurrency::Ctrie<int>::Allocator());
+  const auto hash_trie = std::allocate_shared<Zeni::Concurrency::Ctrie<int, Hash_Mod<int, 10>, std::equal_to<int>, uint8_t>>(Zeni::Concurrency::Ctrie<int>::Allocator());
   for (int i = 0; i != 64; ++i)
     job_queue->give_one(std::make_shared<Ctrie_Job>(hash_trie, 63));
   worker_threads->finish_jobs();
